Moved the clean_ind_cqc job, validation and tests to new folder structure.
