+++ projects/_03_independent_cqc/unittest_data/ind_cqc_test_file_data.py
@@ -411,3 +411,308 @@ class ArchiveFilledPostsEstimates:
     expected_single_digit_number_as_string = "09"
     double_digit_number = 10
     expected_double_digit_number_as_string = "10"
+
+
+@dataclass
+class CleanIndCQCData:
+    # fmt: off
+    merged_rows_for_cleaning_job = [
+        ("1-1000001", "20220201", date(2020, 2, 1), "South East", "Surrey", "Rural", "Y", 0, 5, 82, None, "Care home without nursing", "name", "postcode", date(2022, 1, 1), "2020", "01", "01"),
+        ("1-1000001", "20220101", date(2022, 1, 1), "South East", "Surrey", "Rural", "Y", 5, 5, None, 67, "Care home without nursing", "name", "postcode", date(2022, 1, 1), "2020", "01", "01"),
+        ("1-1000002", "20220101", date(2022, 1, 1), "South East", "Surrey", "Rural", "N", 0, 17, None, None, "non-residential", "name", "postcode", date(2022, 1, 1), "2020", "01", "01"),
+        ("1-1000002", "20220201", date(2022, 2, 1), "South East", "Surrey", "Rural", "N", 0, 34, None, None, "non-residential", "name", "postcode", date(2022, 1, 1), "2020", "01", "01"),
+        ("1-1000003", "20220301", date(2022, 3, 1), "North West", "Bolton", "Urban", "N", 0, 34, None, None, "non-residential", "name", "postcode", date(2022, 1, 1), "2020", "01", "01"),
+        ("1-1000003", "20220308", date(2022, 3, 8), "North West", "Bolton", "Rural", "N", 0, 15, None, None, "non-residential", "name", "postcode", date(2022, 1, 1), "2020", "01", "01"),
+        ("1-1000004", "20220308", date(2022, 3, 8), "South West", "Dorset", "Urban", "Y", 9, 0, 25, 25, "Care home with nursing", "name", "postcode", date(2022, 1, 1), "2020", "01", "01"),
+    ]
+    # fmt: on
+
+    remove_cqc_duplicates_when_carehome_and_asc_data_populated_rows = [
+        (
+            "loc 1",
+            date(2024, 1, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2018, 1, 1),
+        ),
+        (
+            "loc 2",
+            date(2024, 1, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2022, 1, 1),
+        ),
+    ]
+    expected_remove_cqc_duplicates_when_carehome_and_asc_data_populated_rows = [
+        (
+            "loc 1",
+            date(2024, 1, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2018, 1, 1),
+        ),
+    ]
+
+    remove_cqc_duplicates_when_carehome_and_asc_data_missing_on_earlier_reg_date_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            None,
+            None,
+            date(2018, 1, 1),
+        ),
+        (
+            "loc 2",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2022, 1, 1),
+        ),
+    ]
+    expected_remove_cqc_duplicates_when_carehome_and_asc_data_missing_on_earlier_reg_date_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2018, 1, 1),
+        ),
+    ]
+
+    remove_cqc_duplicates_when_carehome_and_asc_data_missing_on_later_reg_date_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2018, 1, 1),
+        ),
+        (
+            "loc 2",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            None,
+            None,
+            date(2022, 1, 1),
+        ),
+    ]
+    expected_remove_cqc_duplicates_when_carehome_and_asc_data_missing_on_later_reg_date_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2018, 1, 1),
+        ),
+    ]
+
+    remove_cqc_duplicates_when_carehome_and_asc_data_missing_on_all_reg_dates_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            None,
+            None,
+            date(2018, 1, 1),
+        ),
+        (
+            "loc 2",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            None,
+            None,
+            date(2022, 1, 1),
+        ),
+    ]
+    expected_remove_cqc_duplicates_when_carehome_and_asc_data_missing_on_all_reg_dates_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            None,
+            None,
+            date(2018, 1, 1),
+        ),
+    ]
+
+    remove_cqc_duplicates_when_carehome_and_asc_data_different_on_all_reg_dates_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2018, 1, 1),
+        ),
+        (
+            "loc 2",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            11,
+            11,
+            date(2022, 1, 1),
+        ),
+    ]
+    expected_remove_cqc_duplicates_when_carehome_and_asc_data_different_on_all_reg_dates_rows = [
+        (
+            "loc 1",
+            date(2024, 2, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2018, 1, 1),
+        ),
+    ]
+
+    remove_cqc_duplicates_when_carehome_and_registration_dates_the_same_rows = [
+        (
+            "loc 1",
+            date(2024, 1, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2022, 1, 1),
+        ),
+        (
+            "loc 1",
+            date(2024, 1, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2022, 1, 1),
+        ),
+    ]
+    expected_remove_cqc_duplicates_when_carehome_and_registration_dates_the_same_rows = [
+        (
+            "loc 1",
+            date(2024, 1, 1),
+            "care home",
+            "AB1 2CD",
+            CareHome.care_home,
+            10,
+            10,
+            date(2022, 1, 1),
+        ),
+    ]
+
+    remove_cqc_duplicates_when_non_res_rows = [
+        (
+            "loc 1",
+            date(2024, 1, 1),
+            "not care home",
+            "AB1 2CD",
+            CareHome.not_care_home,
+            None,
+            None,
+            date(2022, 1, 1),
+        ),
+        (
+            "loc 2",
+            date(2024, 1, 1),
+            "not care home",
+            "AB1 2CD",
+            CareHome.not_care_home,
+            10,
+            10,
+            date(2022, 1, 1),
+        ),
+    ]
+    expected_remove_cqc_duplicates_when_non_res_rows = (
+        remove_cqc_duplicates_when_non_res_rows
+    )
+
+    repeated_value_rows = [
+        ("1", 1, date(2023, 2, 1)),
+        ("1", 2, date(2023, 3, 1)),
+        ("1", 2, date(2023, 4, 1)),
+        ("1", 3, date(2023, 8, 1)),
+        ("2", 3, date(2023, 2, 1)),
+        ("2", 9, date(2023, 4, 1)),
+        ("2", 3, date(2024, 1, 1)),
+        ("2", 3, date(2024, 2, 1)),
+    ]
+
+    expected_without_repeated_values_rows = [
+        ("1", 1, date(2023, 2, 1), 1),
+        ("1", 2, date(2023, 3, 1), 2),
+        ("1", 2, date(2023, 4, 1), None),
+        ("1", 3, date(2023, 8, 1), 3),
+        ("2", 3, date(2023, 2, 1), 3),
+        ("2", 9, date(2023, 4, 1), 9),
+        ("2", 3, date(2024, 1, 1), 3),
+        ("2", 3, date(2024, 2, 1), None),
+    ]
+
+
+@dataclass
+class ValidateCleanedIndCqcData:
+    # fmt: off
+    merged_ind_cqc_rows = [
+        ("1-000000001", date(2024, 1, 1), CareHome.care_home, "name", "postcode", "2024", "01", "01"),
+        ("1-000000002", date(2024, 1, 1), CareHome.not_care_home, "name", "postcode", "2024", "01", "01"),
+        ("1-000000001", date(2024, 2, 1), CareHome.care_home, "name", "postcode", "2024", "02", "01"),
+        ("1-000000002", date(2024, 2, 1), CareHome.not_care_home, "name", "postcode", "2024", "02", "01"),
+    ]
+
+    cleaned_ind_cqc_rows = [
+        ("1-000000001", date(2024, 1, 1), date(2024, 1, 1), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, ["service"], PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", 5, "estab_1", "org_1", 5, 5, "source", 5.0, 5.0, 5, "2024", "01", "01"),
+        ("1-000000002", date(2024, 1, 1), date(2024, 1, 1), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, ["service"], PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", 5, "estab_1", "org_1", 5, 5, "source", 5.0, 5.0, 5, "2024", "01", "01"),
+        ("1-000000001", date(2024, 1, 9), date(2024, 1, 1), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, ["service"], PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", 5, "estab_1", "org_1", 5, 5, "source", 5.0, 5.0, 5, "2024", "01", "09"),
+        ("1-000000002", date(2024, 1, 9), date(2024, 1, 1), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, ["service"], PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", 5, "estab_1", "org_1", 5, 5, "source", 5.0, 5.0, 5, "2024", "01", "09"),
+    ]
+    # fmt: on
+
+    calculate_expected_size_rows = [
+        (
+            "1-000000001",
+            date(2024, 1, 1),
+            CareHome.care_home,
+            "name",
+            "postcode",
+            "2024",
+            "01",
+            "01",
+        ),
+    ]
